- suppressed some errors in printf commands with uint64_t                                                                                               - apparently, the cap function in QRegExp takes a non const in some Qt versions. I had to copy the const parameter to a local variable.
git-svn-id: http://svn.code.sf.net/p/retroshare/code/trunk@2918 b45a01b8-16f6-495d-af2f-9b41ad6348cc

+++ retroshare-gui/src/gui/forums/CreateForumMsg.cpp
@@ -368,7 +368,7 @@ void CreateForumMsg::fileHashingFinished(AttachFileItem* file) {
 
 	//convert fileSize from uint_64 to string for html link
 	char fileSizeChar [100];
-	sprintf(fileSizeChar, "%lld", file->FileSize());
+	sprintf(fileSizeChar, "%llu", (unsigned long long int)file->FileSize());
 	std::string fileSize = *(&fileSizeChar);
 
 	std::string mesgString = "<a href='retroshare://file|" + (file->FileName()) + "|" + fileSize + "|" + (file->FileHash()) + "'>"
